Clean move polyline decoding method to util
