+++ app/src/main/java/io/github/vibrouter/models/DirectionsApiResult.java
@@ -5,9 +5,7 @@ import com.google.android.gms.maps.model.LatLng;
 import java.util.ArrayList;
 import java.util.List;
 
-/**
- * Created by yishihara on 17/01/29.
- */
+import static io.github.vibrouter.utils.DirectionApiUtil.decodePolyline;
 
 public class DirectionsApiResult {
     public static class GeocodedWaypoint {
@@ -93,40 +91,5 @@ public class DirectionsApiResult {
                 Double.valueOf(endLocation.lng));
         waypoints.add(endLatLng);
         return waypoints;
-    }
-
-    public static List<LatLng> decodePolyline(String encoded) {
-        List<LatLng> poly = new ArrayList<>();
-        int index = 0;
-        int lat = 0;
-        int lng = 0;
-
-        while (index < encoded.length()) {
-            int b, shift = 0, result = 0;
-            do {
-                b = encoded.charAt(index++) - 63;
-                result |= (b & 0x1f) << shift;
-                shift += 5;
-            } while (b >= 0x20);
-
-            int dlat = ((result & 1) != 0 ? ~(result >> 1) : (result >> 1));
-            lat += dlat;
-
-            shift = 0;
-            result = 0;
-            do {
-                b = encoded.charAt(index++) - 63;
-                result |= (b & 0x1f) << shift;
-                shift += 5;
-            } while (b >= 0x20);
-
-            int dlng = ((result & 1) != 0 ? ~(result >> 1) : (result >> 1));
-            lng += dlng;
-
-            LatLng p = new LatLng((((double) lat / 1E5)), (((double) lng / 1E5)));
-            poly.add(p);
-        }
-
-        return poly;
     }
 }
